Moved several component files and updated import paths in page.js
